refactor: create token specific utils file, move to shared
