+++ src/shared/utils/index.ts
@@ -25,80 +25,6 @@ export const delay = (delayMs: number) => {
   )
 }
 
-const tiniestAmount = 0.000001
-
-const formatWithDigits = (number: string, resultDecimals?: number) => {
-  const splitArr = number.split('.')
-  const secondPart =
-    splitArr[1].length > 8
-      ? splitArr[1].slice(0, resultDecimals || 9)
-      : splitArr[1]
-
-  return [splitArr[0], secondPart].join('.')
-}
-
-export const formatSmallNumbers = (
-  smallNumber: string | number,
-  resultDecimals?: number,
-) => {
-  if (isNaN(Number(smallNumber))) {
-    return smallNumber.toString()
-  }
-
-  smallNumber = smallNumber.toString()
-  const asNumber = Number(smallNumber)
-
-  if (asNumber >= tiniestAmount) {
-    return formatWithDigits(smallNumber, resultDecimals)
-  }
-
-  return asNumber !== 0 ? `< ${tiniestAmount}` : '0.00'
-}
-
-// this needs to be here because of the failing tests
-enum TokenSymbol {
-  TRBTC = 'TRBTC',
-  RBTC = 'RBTC',
-}
-
-export const rbtcMap = new Map([
-  [TokenSymbol.TRBTC, true],
-  [TokenSymbol.RBTC, true],
-  [undefined, false],
-])
-
-export const formatTokenValues = (
-  number: string | number,
-  resultDecimals?: number,
-) => {
-  // make sure to use this only at the end when showing values
-  if (isNaN(Number(number))) {
-    return number.toString()
-  }
-  number = number.toString()
-  const asNumber = Number(number)
-
-  if (asNumber < 1) {
-    return formatSmallNumbers(number, resultDecimals)
-  }
-
-  if (number.includes('.') && asNumber > 1) {
-    return formatWithDigits(number, resultDecimals)
-  }
-
-  if (number.length <= 3) {
-    return number
-  }
-
-  const longNumberArr = number.split('')
-
-  for (let i = number.length - 3; i > 0; i -= 3) {
-    longNumberArr.splice(i, 0, ',')
-  }
-
-  return longNumberArr.join('')
-}
-
 export const errorHandler = (error: unknown) => {
   if (typeof error === 'object' && Object.hasOwn(error as object, 'message')) {
     const err = error as ErrorWithMessage
@@ -161,41 +87,5 @@ export const usePreventScreenshot = (
   }, [isFocused])
 }
 
-export const getFormattedTokenValue = (tokenValue: string) => {
-  if (!tokenValue.includes('.')) {
-    return tokenValue
-  }
-  const tokenValueArr = tokenValue.split('.')
-  const decimals = tokenValueArr[1].length
-
-  if (decimals < 8) {
-    return tokenValue
-  }
-
-  const restDecimals = tokenValueArr[1].split('').slice(7, decimals)
-
-  let moreThanZeroIndex = 0
-
-  for (let i = 0; i < restDecimals.length; i++) {
-    if (Number(restDecimals[i]) > 0) {
-      moreThanZeroIndex = i
-      break
-    }
-  }
-
-  const hasOneMoreDigit = !!restDecimals[moreThanZeroIndex + 1]
-  const lastAfterZero = restDecimals.slice(
-    0,
-    hasOneMoreDigit ? moreThanZeroIndex + 2 : moreThanZeroIndex + 1,
-  )
-  const ending = restDecimals[moreThanZeroIndex + 2] ? '...' : ''
-
-  return (
-    tokenValueArr[0] +
-    '.' +
-    tokenValueArr[1].slice(0, 7).concat(...lastAfterZero) +
-    ending
-  )
-}
-
+export * from './tokenValues'
 export * from './wallet'
