+++ src/subscriptions/rifSockets.ts
@@ -9,13 +9,14 @@ import { ChainID } from 'lib/eoaWallet'
 
 import { resetSocketState } from 'store/shared/actions/resetSocketState'
 import { AppDispatch } from 'store/index'
-import { abiEnhancer, getDefaultTokens } from 'core/setup'
+import { abiEnhancer } from 'core/setup'
 import { addOrUpdateBalances } from 'store/slices/balancesSlice'
 import { TokenBalanceObject } from 'store/slices/balancesSlice/types'
 import { UsdPricesState } from 'store/slices/usdPricesSlice'
 import { getWalletSetting } from 'core/config'
 import { SETTINGS } from 'core/types'
 import { MMKVStorage } from 'storage/MMKVStorage'
+import { getDefaultTokens } from 'shared/utils'
 
 import { onSocketChangeEmitted } from './onSocketChangeEmitted'
 import { Action, InitAction } from './types'
